tool_project/pmt-backend2/src/main/java/com/example/pmt_backend/DTO/RoleDTO.java
package com.example.pmt_backend.DTO;



public class RoleDTO {
    private Long projectId;
    private Long userId;
    private String role;

    // Getters and Setters

    public Long getProjectId() {
        return projectId;
    }

    public void setProjectId(Long projectId) {
        this.projectId = projectId;
    }

    public Long getUserId() {
        return userId;
    }

    public void setUserId(Long userId) {
        this.userId = userId;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }
}
